backend/requirements.txt
# Core
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.1
pydantic-settings==2.8.1

# Database
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.2
greenlet==3.2.1

# Redis
redis[hiredis]==5.2.1

# HTTP
httpx==0.28.1

# Security
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0

# Utilities
python-multipart==0.0.20
python-dotenv==1.1.0
orjson==3.10.16
